Comparison Class

The `Comparison` class represents a comparison between two groups of samples. It includes methods to initialize, print, and validate the data.

Details

An R6 class that represents a comparison between two groups of samples. This class contains the comparison name, the group order, and the comparison table.

Initialization

``` Comparison$new( comparison_name = "My comparison", group_order = c("Control group", "Test group"), comparison_table = data.table::data.table( group = c("Control group", "Control group", "Test group", "Test group"), sample = c("Sample1", "Sample2", "Sample3", "Sample4") ) ) ```

Usage

``` comparison <- Comparison$new( comparison_name = "My comparison", group_order = c("Control group", "Test group"), comparison_table = data.table::data.table( group = c("Control group", "Control group", "Test group", "Test group"), sample = c("Sample1", "Sample2", "Sample3", "Sample4") ) ) comparison$print() # Print the Comparison object ```

Public fields

comparison_name

Character. The name of the comparison.

group_order

Character vector. The order of groups for the comparison, with length 2. The first element will be treated as the "control" group, and the second as the "test" group.

comparison_table

A data.table that contains the group and sample information for the comparison.

Methods

Public methods

Comparison$new()

Comparison$print()

Comparison$clone()

Method new()

Usage

Comparison$new(comparison_name, group_order, comparison_table)

Arguments

comparison_name

A character string representing the name of the comparison. Must be of length 1 and not exceed 100 characters.

group_order

A character vector specifying the order of groups for the comparison. Must be of length 2.

comparison_table

A data.table that contains group and sample information for the comparison. This table is crucial for tracking which samples belong to which clinical group. It is used to distinguish samples for the purposes of experimental analysis. It should have two columns, named "group" and "sample". The "group" column identifies the clinical group to which each sample belongs, and the "sample" column lists the names/IDs of the samples. The groups in this table should match the names specified in `group_order`.

Method print()

Usage

Comparison$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

Comparison$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.