Candlesticks financial `ggplot2` layer

Candlesticks financial `ggplot2` layer

Usage

stat_candlestick(
  mapping = NULL,
  data = NULL,
  geom = "linerange",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  colours = list(up = "#55BE8B", down = "#ED4D5D", no_change = "#535453"),
  ...
)

Arguments

mapping

A `ggplot2::aes` object (required - default `NULL`).

`x`: The x-axis value, usually representing time.

`open`: The opening price.

`close`: The closing price.

`high`: The highest price in the time range.

`low`: The lowest price in the time range.

`group`: (optional) The grouping variable.

data

A `data.table` object (required - default `NULL`).

colours

A `list` with three elements "up", "down", and "no_change". These are the colours of the candlesticks when a positive change in price action, a negative change and no change respectively.

...

Additional arguments passed to `ggplot2::layer`.

Value

A `ggplot2::layer` object.

Details

This is a `ggplot2` extension; it is used with the `+` operator for adding a layer to a `ggplot2` object.

Author

Dereck Mezquita

Examples

# get some financial data
# kucoin is a private package - you can use any data source
ticker <- "BTC/USDT"

dt <- kucoin::get_market_data(
    symbols = ticker,
    from = "2022-11-28 15:29:43 EST", # lubridate::now() - lubridate::days(7),
    to = "2022-12-05 15:29:31 EST",# lubridate::now(),
    frequency = "1 hour"
)
#> Error in loadNamespace(x): there is no package called ‘kucoin’

dt
#> function (x, df, ncp, log = FALSE) 
#> {
#>     if (missing(ncp)) 
#>         .Call(C_dt, x, df, log)
#>     else .Call(C_dnt, x, df, ncp, log)
#> }
#> <bytecode: 0x560fcdfa9580>
#> <environment: namespace:stats>

dt |>
    ggplot2::ggplot(ggplot2::aes(
        x = datetime,
        open = open,
        close = close,
        high = high,
        low = low,
        group = symbol
    )) +
    ## ------------------------------------
    dmplot::stat_candlestick() +
    ## ------------------------------------
    ggplot2::scale_x_continuous(n.breaks = 25, labels = \(x) {
        lubridate::floor_date(lubridate::as_datetime(x), "hours")
    }) +
    ggplot2::scale_y_continuous(n.breaks = 25) +
    ggplot2::labs(
        title = ticker,
        x = "Date",
        y = "Price (USD)"
    ) +
    dmplot::theme_dereck_dark() +
    ggplot2::theme(
        axis.text.x = ggplot2::element_text(angle = 75, vjust = 0.925, hjust = 0.975),
        panel.grid.minor = ggplot2::element_blank()
    )
#> Error in ggplot2::ggplot(dt, ggplot2::aes(x = datetime, open = open, close = close,     high = high, low = low, group = symbol)): `data` cannot be a function.
#> ℹ Have you misspelled the `data` argument in `ggplot()`