Moving average convergence divergence (macd) `ggplot2` layer

`stat_macd` is a `ggplot2` layer that allows you to plot a `ggplot2` layer of moving average convergence divergence (macd) by providing the column names `ggplot2::aes` of the previously calculated metrics.

You are free to use whatever algorithm you desire; the result will be a line plot overlayed on a histogram.

To use this layer provide `ggplot2::aes` values for `x` (datetime x-axis) and `macd` (y-axis) and `macd_signal` (y-axis) and `macd_diff` (y-axis).

See examples for more details.

Usage

stat_macd(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  alpha_lines = 0.75,
  alpha_histogram = 0.5,
  colour_lines = list(macd = "blue", macd_signal = "red"),
  colour_histogram = list(positive = NA_character_, negative = NA_character_),
  fill_histogram = list(positive = "green", negative = "red"),
  ...
)

Arguments

mapping

A `ggplot2::aes` object (required - default `NULL`).

data

A `data.table` object (required - default `NULL`).

alpha_lines

A `numeric` vector of length one; the alpha of the lines (optional - default `0.75`).

alpha_histogram

A `numeric` vector of length one; the alpha of the histogram (optional - default `0.5`).

colour_lines

A named or unnamed `list` of two elements "macd" and "macd_signal" (optional - default `list(macd = "blue", macd_signal = "red")`).

colour_histogram

A named or unnamed `list` of two elements "positive" and "negative" (optional - default `list(positive = NA_character_, negative = NA_character_)`).

fill_histogram

A named or unnamed `list` of two elements "positive" and "negative" (optional - default `list(positive = "green", negative = "red")`).

linewidth

A `numeric` vector of length one; the linewidth of the line (optional - default `1.75`).

Value

A `ggplot2::layer` object.

Details

This is a `ggplot2` extension; it is used with the `+` operator for adding a layer to a `ggplot2` object.

Aesthetics

stat_movingaverages understands the following aesthetics (required aesthetics are in bold):

x – datetime (x-axis)

macd – y-axis

macd_signal – y-axis

macd_diff – y-axis

Author

Dereck Mezquita

Examples

# get some financial data
# kucoin is private package - you can use any data source
ticker <- "BTC/USDT"

dt <- kucoin::get_market_data(
    symbols = ticker,
    from = "2022-11-28 15:29:43 EST", # lubridate::now() - lubridate::days(7),
    to = "2022-12-05 15:29:31 EST",# lubridate::now(),
    frequency = "1 hour"
)
#> Error in loadNamespace(x): there is no package called ‘kucoin’

dt
#> function (x, df, ncp, log = FALSE) 
#> {
#>     if (missing(ncp)) 
#>         .Call(C_dt, x, df, log)
#>     else .Call(C_dnt, x, df, ncp, log)
#> }
#> <bytecode: 0x5590f94e8a00>
#> <environment: namespace:stats>

# we need a function that calculates the indicator for us
# typically I like to write my own functions in C++; in this case we will use TTR's
# the stat expects a named list to be returned - we redefine ttr
macd <- function(x, s = 12, l = 26, k = 9) {as.list(as.data.frame(TTR::MACD(x, s, l, k)))}

# macd(dt$close)
dt[, c("macd", "macd_signal") := macd(close, s = 12, l = 26, k = 9)]
#> Error in `:=`(c("macd", "macd_signal"), macd(close, s = 12, l = 26, k = 9)): could not find function ":="
dt[, macd_diff := macd - macd_signal]
#> Error in `:=`(macd_diff, macd - macd_signal): could not find function ":="

na.omit(dt) |>
    ggplot2::ggplot(ggplot2::aes(x = datetime)) +
    dmplot::stat_macd(ggplot2::aes(macd = macd, macd_signal = macd_signal, macd_diff = macd_diff)) +
    ## ------------------------------------
    # provide the colnames to the calculated indicators as aes values
    ## ------------------------------------
    ggplot2::scale_x_continuous(n.breaks = 25, labels = \(x) {
        lubridate::floor_date(lubridate::as_datetime(x), "hours")
    }) +
    ggplot2::scale_y_continuous(n.breaks = 25) +
    ggplot2::labs(
        title = ticker,
        x = "Date",
        y = "Price (USD)"
    ) +
    dmplot::theme_dereck_dark() +
    ggplot2::theme(
        axis.text.x = ggplot2::element_text(angle = 75, vjust = 0.925, hjust = 0.975),
        panel.grid.minor = ggplot2::element_blank()
    )
#> Error in ggplot2::ggplot(na.omit(dt), ggplot2::aes(x = datetime)): `data` cannot be a function.
#> ℹ Have you misspelled the `data` argument in `ggplot()`