Moving averages `ggplot2` layer

`stat_movingaverages` is a `ggplot2` layer that allows you to plot moving averages on a `ggplot2` plot either by providing the column names `ggplot2::aes` of the previously calculated metrics.

You are free to use whatever algorithm you desire; the result will be two line plots one for a short moving average and one for a long moving average.

To use this layer provide `ggplot2::aes` values for `x` (datetime x-axis) and `short` and `long` (y-axis).

See examples for more details.

Usage

stat_movingaverages(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  linewidth = 1.75,
  alpha = 0.75,
  colour = list(short = "yellow", long = "purple"),
  ...
)

Arguments

mapping

A `ggplot2::aes` object (required - default `NULL`).

data

A `data.table` object (required - default `NULL`).

linewidth

A `numeric` vector of length one; the width of the line (optional - default `1.75`).

alpha

A `numeric` vector of length one; the alpha of the line (optional - default `0.75`).

colour

A named or unnamed `list` with three elements "short" and "long". These are the colours for the short and long moving averages (optional - default `list(short = "red", long = "blue")`).

Value

A `ggplot2::layer` object.

Details

This is a `ggplot2` extension; it is used with the `+` operator for adding a layer to a `ggplot2` object.

Aesthetics

stat_movingaverages understands the following aesthetics (required aesthetics are in bold):

x – datetime (x-axis)

short – the values for the short moving average (y-axis).

long – the values for the long moving average (y-axis)

Author

Dereck Mezquita

Examples

# get some financial data
# kucoin is private package - you can use any data source
ticker <- "BTC/USDT"

dt <- kucoin::get_market_data(
    symbols = ticker,
    from = "2022-11-28 15:29:43 EST", # lubridate::now() - lubridate::days(7),
    to = "2022-12-05 15:29:31 EST", # lubridate::now(),
    frequency = "1 hour"
)
#> Error in loadNamespace(x): there is no package called ‘kucoin’

dt
#> function (x, df, ncp, log = FALSE) 
#> {
#>     if (missing(ncp)) 
#>         .Call(C_dt, x, df, log)
#>     else .Call(C_dnt, x, df, ncp, log)
#> }
#> <bytecode: 0x5652a9bc3418>
#> <environment: namespace:stats>

# we need a function that calculates the indicator for us
# typically I like to write my own functions in C++; in this case we will use TTR's
# the stat expects a named list to be returned - we redefine ttr
ema <- function(close, n = 2, wilder = TRUE) {
    return(as.list(as.data.frame(TTR::EMA(close, n = n, wilder = wilder))))
}

# calculate the short and long moving averages
dt[, ema_short := ema(close, n = 10, wilder = TRUE)]
#> Error in `:=`(ema_short, ema(close, n = 10, wilder = TRUE)): could not find function ":="
dt[, ema_long := ema(close, n = 50, wilder = TRUE)]
#> Error in `:=`(ema_long, ema(close, n = 50, wilder = TRUE)): could not find function ":="

dt |>
    ggplot2::ggplot(ggplot2::aes(
        x = datetime,
        open = open,
        close = close,
        high = high,
        low = low,
        group = symbol
    )) +
    ## ------------------------------------
    dmplot::stat_candlestick() +
    ## ------------------------------------
    # provide the colnames to the calculated indicators as aes values
    dmplot::stat_movingaverages(ggplot2::aes(short = ema_short, long = ema_long), alpha = list(mavg = 0.5)) +
    ## ------------------------------------
    ggplot2::scale_x_continuous(n.breaks = 25, labels = \(x) {
        lubridate::floor_date(lubridate::as_datetime(x), "hours")
    }) +
    ggplot2::scale_y_continuous(n.breaks = 25) +
    ggplot2::labs(
        title = ticker,
        x = "Date",
        y = "Price (USD)"
    ) +
    dmplot::theme_dereck_dark() +
    ggplot2::theme(
        axis.text.x = ggplot2::element_text(angle = 75, vjust = 0.925, hjust = 0.975),
        panel.grid.minor = ggplot2::element_blank()
    )
#> Error in ggplot2::ggplot(dt, ggplot2::aes(x = datetime, open = open, close = close,     high = high, low = low, group = symbol)): `data` cannot be a function.
#> ℹ Have you misspelled the `data` argument in `ggplot()`